The great 64-bit rewrite * We did loose tasking and only have a paging skeleton, but hey it took like 4 months so i'm happy that there's something :p

+++ kernel/include/pmm.h
@@ -7,7 +7,7 @@
 extern bitmap_t pmm_bitmap;
 
 int pmm_init(void);
-uint32_t pmm_alloc(void);
-void pmm_free(uint32_t page);
+size_t pmm_alloc(void);
+void pmm_free(size_t page);
 
 #endif
